+++ CHANGELOG.md
@@ -33,7 +33,7 @@ The general format is:
 
 ### Added
 
-- add support for attaching a secret volume to Workflow
+- add support for attaching a secret volume to Workflows and CronWorkflows
 # 1.3.0 - DATE (20/01/2022)
 
 ### Added